+++ CHANGELOG.md
@@ -20,6 +20,9 @@ functionality, under the "Removed" section.
 
 - Passing the `--bypass-root-check` flag now elevates if the user is not root
   instead of calling Nix as non-root user.
+- A new `nh os build-image` subcommand for building a disk-image variant is now
+  available. A variant can be selected using the `--image-variant` flag. All
+  variants in the `config.system.build.images` attribute set are supported.
 - `--elevation-program` flag was renamed to `--elevation-strategy` with support
   for `'none'` (no elevation) and `'passwordless'` (for remote hosts with
   `NOPASSWD` configured) values. The old flag name remains available as an alias